Crate minion

This crate provides ability to spawn highly isolated processes

Platform support

warning: not all features are supported by all backends. See documentation for particular backend to know more

Re-exports

pub use crate::linux::LinuxBackend;

pub use crate::linux::LinuxChildProcess;

pub use linux::LinuxHandle;

Modules

command

errors

linux

Structs

ChildProcessOptions
This type should only be used by Backend implementations Use Command instead

Command
Child process builder

DominionOptions

DominionRef
Type-erased dominion

InputSpecification

LinuxDominion

OutputSpecification

PathExpositionOptions

ResourceUsageData
This struct is returned by Dominion::query_usage_data It represents various resource usage Some items can be absent or rounded

StdioSpecification
Specifies how to provide child stdio

Enums

DesiredAccess
Mount options.

DominionRefInner

Error

InputSpecificationData
Configures stdin for child

OutputSpecificationData
Configures stdout and stderr for child

WaitOutcome
Returned by ChildProcess::wait_for_exit

Traits

Backend
Represents way of isolation

ChildProcess
Represents child process.

Dominion
Represents highly-isolated sandbox

Functions

check
This functions checks for system configurations issues. If it returns None, minion will probably work. If it returns Some(s), s is human-readable string describing these problems. It should be shown to administrtor, so that they can fix this problem.

setup

Type Definitions

DefaultBackend

Result