Trait minion::ChildProcess

pub trait ChildProcess: Debug {
    fn get_exit_code(&self) -> Result<Option<i64>>;

    fn stdin(&mut self) -> Option<Box<dyn Write + Send + Sync>>;

    fn stdout(&mut self) -> Option<Box<dyn Read + Send + Sync>>;

    fn stderr(&mut self) -> Option<Box<dyn Read + Send + Sync>>;

    fn wait_for_exit(&self, timeout: Option<Duration>) -> Result<WaitOutcome>;

    fn poll(&self) -> Result<()>;

    fn is_finished(&self) -> Result<bool>;
}

Represents child process.

Required methods

fn get_exit_code(&self) -> Result<Option<i64>>

Returns exit code, if process had exited by the moment of call, or None otherwise.

fn stdin(&mut self) -> Option<Box<dyn Write + Send + Sync>>

Returns writeable stream, connected to child stdin

Stream will only be returned, if corresponding Stdio item was new_pipe. Otherwise, None will be returned

On all subsequent calls, None will be returned

fn stdout(&mut self) -> Option<Box<dyn Read + Send + Sync>>

Returns readable stream, connected to child stdoutn

Stream will only be returned, if corresponding Stdio item was new_pipe. Otherwise, None will be returned

On all subsequent calls, None will be returned

fn stderr(&mut self) -> Option<Box<dyn Read + Send + Sync>>

Returns readable stream, connected to child stderr

Stream will only be returned, if corresponding Stdio item was new_pipe. Otherwise, None will be returned

On all subsequent calls, None will be returned

fn wait_for_exit(&self, timeout: Option<Duration>) -> Result<WaitOutcome>

Waits for child process exit with timeout. If timeout is None, wait_for_exit will block until child has exited

fn poll(&self) -> Result<()>

Refreshes information about process

fn is_finished(&self) -> Result<bool>

Returns whether child process has exited by the moment of call This function doesn't blocks on waiting (see wait_for_exit).

Loading content...